refactor: move configs to index.ts
